Split code into multiple files
