+++ server/database/environment.go
@@ -1,0 +1,42 @@
+package database
+
+import (
+	"database/sql"
+	"fmt"
+
+	_ "github.com/lib/pq"
+)
+
+const (
+	host     = "localhost"
+	port     = 5432
+	user     = "dev"
+	password = "hunter2"
+	dbname   = "feelings"
+)
+
+type Env struct {
+	db *sql.DB
+}
+
+func OpenDbEnv() *Env {
+	psqlInfo := fmt.Sprintf("host=%s port=%d user=%s "+
+		"password=%s dbname=%s sslmode=disable",
+		host, port, user, password, dbname)
+	db, err := sql.Open("postgres", psqlInfo)
+	if err != nil {
+		panic(err)
+	}
+
+	err = db.Ping()
+	if err != nil {
+		panic(err)
+	}
+
+	fmt.Println("Successfully connected!")
+	return &Env{db: db}
+}
+
+func (e *Env) Close() {
+	e.db.Close()
+}
